Added optional FAST_JX/v2024-05-Hg/ folder
Due to the updates in PR 
https://github.com/geoschem/geos-chem/pull/2238,
we also need to add the updated org.dat and so4.dat files to
the FAST_JX folder for the Hg simulation.  Created a new folder
CHEM_INPUTS/FAST_JX/v2024-05-Hg with these new files and
the previous files from v2022-10-Hg.  This is specified as an
optional folder.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>